requirements.txt
Django>=3.0,<4.0
django-import-export
django-crispy-forms
crispy-tailwind
django-htmx
xlwt
